==> Analysis Code/2c Stepwise Regressions for Z vars.Rmd <==
---
title: "Running Regressions: Step wise Regressions for Identifying the Z vars"
output:
  html_notebook:
    number_sections: yes
  html_document: default
  pdf_document: default
date: '`r Sys.Date()`'
---
##Summary
This file uses a sort of adapted step wise regression to identify the best spec for each outcome. I initially start with a specification that involves all the market terms (triple interactions), each iteration eliminates the term with the highest p value. 

The initial iteration however, did not go that well, since with the threshold of ).2 for P val plan was to do this without consideration for whether the term is a triple interaction with GQ treatment or NSEW treatment. However, this led to several outcomes without any terms at all. As a solution, I have added an additional condition that checks to see the number of market terms for each treatment category.

In addition, once we hit the threshold p value for a particular treatment (GQ/NSEW), I stop removing terms from them, eventhough subsequent iterations were terms from the other treatment are removed might lead to a new term being above the threshold.

##Data Loading and Preparation
```{r, message=FALSE}
rm(list = ls())
library(tidyverse); library(plm); library(broom); library(stringr)
##The regression data
load("../data/1 Cleaned files for analysis/allData.RDA")
        
##The list of variables
yVars <- read_csv("../data/1 Cleaned files for analysis/Regression Variables/yVarsFinal.csv") %>% 
        select(varNames, yType, description) %>% 
        arrange(yType, varNames)

zVars <- read_csv("../data/1 Cleaned files for analysis/Regression Variables/zVars.csv") %>% 
        arrange(zType, varNames)

```


##Functions to run regression
Construct a formula given a list of Z (market) variables and a Y (outcome) variable.
```{r}
createFormula <- function(yVariable, gqVars, nsewVars){
        gqMarketTerms <- paste("gqDistType * postGQ *", gqVars)
        nsewMarketTerms <- paste("nsewDistType * postNSEW *", nsewVars)
        
        allTerms <- paste("gqDistType * postGQ", "nsewDistType * postNSEW", paste(gqMarketTerms, collapse =" + "), paste(nsewMarketTerms, collapse = " + "), "spatialState * year", sep = " + ")
        
        as.formula(paste(yVariable, allTerms, sep = " ~ ")) ##output the formula
        
}

```

This section identifies the final specifications based on the p value criteria. I had to tweak the discussed settings slightly to make sure that all the Z variables were not dropped. The threshold for p value is set as 0.3 and the additional condition forces the loop to exit if we remove more than 9 (out of the 11) terms for either GQ or NSEW.


```{r}
# ##Initiate all the variables for the regression
zVarsGQ <- zVars$varNames 
zVarsNSEW <- zVars$varNames
regressionSpecs <- character()
runCharacteristics <- tibble(yVar = character(),gqRemoved = numeric(), nsewRemoved = numeric(), minPValGQ = numeric(), minPvalNSEW = numeric())
plim <- 0.3
lengthLim <- 5

for(j in 1:length(yVars$varNames)){
        y <- yVars$varNames[j]
        removeTermsGQ <- character() ##set as empty to begin loop
        removeTermsNSEW <- character()
        maxPValGQ <- 1
        maxPValNSEW <- 1
        while(!(maxPValGQ < plim & maxPValNSEW < plim) & (length(removeTermsGQ) < lengthLim | length(removeTermsNSEW) <  lengthLim)){
                tempTibble <- tibble(yVar = y, gqRemoved = length(removeTermsGQ), nsewRemoved = length(removeTermsNSEW), minPValGQ = maxPValGQ, minPvalNSEW = maxPValNSEW)
                runCharacteristics <- bind_rows(runCharacteristics, tempTibble)
                
                # print(c(maxPValGQ, maxPValNSEW))
                # print(c(removeTermsGQ))
                # print(removeTermsNSEW)
                
                ##Create new formula based on remove terms
                regressionFormula <- createFormula(yVariable = y, gqVars = zVarsGQ[!(zVarsGQ %in% removeTermsGQ)], nsewVars = zVarsNSEW[!(zVarsNSEW %in% removeTermsNSEW)])
                
                ##run the regression
                model <- plm(regressionFormula, data = allData)
                
                ##Patterns for extracting GQ and NSEW estimates
                patternGQ <- "(?=.*gqDistType0-40)(?=.*postGQ(?!$))"
                patternNSEW <- "(?=.*nsewDistType0-40)(?=.*postNSEW(?!$))"
                
                ##Extract the estimates with max p value
                gqEstimates <- tidy(model) %>% 
                        filter(str_detect(term, patternGQ)) %>% 
                        filter(p.value == max(p.value, na.rm = T))
                
                nsewEstimates <- tidy(model) %>% 
                        filter(str_detect(term, patternNSEW)) %>% 
                        filter(p.value == max(p.value, na.rm = T))
                
                ##Identify max pvalue and terms to be removed from formula (if current value is greater than threshold)
                maxPValGQ <- gqEstimates$p.value#ifelse(maxPValGQ > plim, gqEstimates$p.value, maxPValGQ) 
                maxPValNSEW <- nsewEstimates$p.value#ifelse(maxPValNSEW > plim, nsewEstimates$p.value, maxPValNSEW)
                
                if(length(removeTermsGQ) > lengthLim){
                        if(length(removeTermsNSEW) > lengthLim){
                                break
                        } 
                        else if(maxPValNSEW > plim){
                              removeTermsNSEW <- c(removeTermsNSEW, str_split(nsewEstimates$term, ":", simplify = T)[1, 3])  
                        } 
                        else{
                                break
                        }
                } 
                else if(maxPValGQ > maxPValNSEW){
                        if(maxPValGQ > plim){
                                removeTermsGQ <- c(removeTermsGQ, str_split(gqEstimates$term, ":", simplify = T)[1, 3])
                        } 
                        else{
                                break
                        }
                }
                else if(length(removeTermsNSEW) > lengthLim){
                        break
                } 
                else if(maxPValNSEW > plim){
                        removeTermsNSEW <- c(removeTermsNSEW, str_split(nsewEstimates$term, ":", simplify = T)[1, 3])  
                } 
                else{
                        break
                }
        }
        regressionSpecs <- c(regressionSpecs, (as.character(regressionFormula)[3])) ##store the terms
}


regressionFormulaFinal <- tibble(yVar = yVars$varNames, terms = regressionSpecs) %>%
        write_csv(., "regSpecs20Apr28OnlyInitial.csv")
```


Creating the table
```{r}
namesList <- as.tibble(expand.grid(highway = c("GQ", "NSEW"), zVars = zVars$varNames)) %>% 
        mutate(termNames = paste(highway, zVars, sep = ":")) %>% 
        select(termNames)

namesList <- bind_rows(tibble(termNames = c("GQ", "NSEW")), namesList) %>% 
        arrange(termNames)
        
```


```{r, message=FALSE}
regressionFormulaFinal <- read_csv("regSpecs20Apr28OnlyInitial.csv") %>% 
        mutate(spec = paste(yVar, terms, sep = " ~ "))

pattern <- "(?=.*DistType0-40)(?=.*post)" ##detect all double and triple interaction terms

runFinal <- function(y){
        yVar <- regressionFormulaFinal[regressionFormulaFinal$spec == y,]$yVar
        model <- plm(as.formula(y), data = allData)
        
        modelDF <- tidy(model) %>% 
                filter(str_detect(term, pattern)) %>%
                mutate(adjRSq = glance(model)$adj.r.squared, nobs = nobs(model)) %>%
                select(term, estimate, p.value, adjRSq, nobs) %>% 
                mutate(term = str_replace_all(term, c("gqDistType0-40:postGQ" = "GQ", "nsewDistType0-40:postNSEW" = "NSEW")))
        
        left_join(namesList, modelDF, by = c("termNames" = "term")) %>% ##join to the namesList
                mutate(yVar = yVar)
        
}


finalResults <- lapply(regressionFormulaFinal$spec, runFinal)

finalResultsDF <- bind_rows(lapply(finalResults, bind_rows)) %>% 
        mutate(estimate = ifelse(is.na(estimate), NA, ifelse(p.value <= 0.01, paste(round(estimate, 4), "***", sep=""), ifelse(p.value <= 0.05, paste(round(estimate, 4), "**", sep=""), ifelse(p.value <= 0.1, paste(round(estimate, 4), "*", sep=""), as.character(round(estimate, 4))))))) 

finalResultsModelChars <- finalResultsDF %>% 
        select(yVar, nobs, adjRSq) %>% 
        group_by(yVar) %>% 
        summarise(nobs = unique(nobs[!is.na(nobs)]), adjRSq = unique(adjRSq[!is.na(adjRSq)])) %>% 
        ungroup()

finalResultsDF <- finalResultsDF %>% 
        select(yVar, term = termNames, estimate) %>% 
        spread(term, estimate) %>% 
        left_join(., finalResultsModelChars, by = "yVar") %>% 
        select(yVar, nobs, adjRSq, one_of(namesList$termNames)) %>% 
        left_join(., yVars, by = c("yVar" = "varNames")) %>% 
        select(1, 22, 23, 2:21) %>% 
        arrange(yType, yVar) %>% 
        mutate(adjRSq = round(adjRSq, 1)) %>% 
        write_csv(., "../Results/Tables/resultsPVal20Apr28OnlyInitial.csv")


```

##Temporary regs

```{r}
specs <- read_csv("regSpecs20Apr26OnlyInitial.csv")
pattern <- "(?=.*DistType0-40)(?=.*post)" ##detect all double and triple interaction terms

specsTotFarm <- specs %>% 
        filter(yVar == "totFarmEmp_f")

formulaTotFarm <- paste("totFarmEmp_tot", specsTotFarm$terms, sep = " ~ ")

model <- plm(as.formula(formulaTotFarm), data = allData)
tidy(model) %>% 
                filter(str_detect(term, pattern)) %>%
                mutate(adjRSq = glance(model)$adj.r.squared, nobs = nobs(model)) %>%
                select(term, estimate, p.value, adjRSq, nobs) %>% 
                mutate(term = str_replace_all(term, c("gqDistType0-40:postGQ" = "GQ", "nsewDistType0-40:postNSEW" = "NSEW"))) %>% 
                mutate(estimate = ifelse(is.na(estimate), NA, ifelse(p.value <= 0.01, paste(round(estimate, 4), "***", sep=""), ifelse(p.value <= 0.05, paste(round(estimate, 4), "**", sep=""), ifelse(p.value <= 0.1, paste(round(estimate, 4), "*", sep=""), as.character(round(estimate, 4)))))))
```

```{r}
specsTotFarm <- specs %>% 
        filter(yVar == "pct_regWage_f_tot")

formulaTotFarm <- paste("pct_regWage_both_tot", specsTotFarm$terms, sep = " ~ ")

model <- plm(as.formula(formulaTotFarm), data = allData)
tidy(model) %>% 
                filter(str_detect(term, pattern)) %>%
                mutate(adjRSq = glance(model)$adj.r.squared, nobs = nobs(model)) %>%
                select(term, estimate, p.value, adjRSq, nobs) %>% 
                mutate(term = str_replace_all(term, c("gqDistType0-40:postGQ" = "GQ", "nsewDistType0-40:postNSEW" = "NSEW"))) %>% 
                mutate(estimate = ifelse(is.na(estimate), NA, ifelse(p.value <= 0.01, paste(round(estimate, 4), "***", sep=""), ifelse(p.value <= 0.05, paste(round(estimate, 4), "**", sep=""), ifelse(p.value <= 0.1, paste(round(estimate, 4), "*", sep=""), as.character(round(estimate, 4)))))))
```



<!-- Given a formula run the regression and update the terms that need to be removed before next iteration -->
<!-- ```{r} -->
<!-- # ##Initiate all the variables for the regression -->
<!-- zVarsGQ <- zVars$varNames  -->
<!-- zVarsNSEW <- zVars$varNames -->
<!-- regressionSpecs <- character() -->
<!-- #length = length(yVars$varNames) -->
<!-- plim <- 0.3 -->

<!-- for(j in 1:length(yVars$varNames)){ -->
<!--         y <- yVars$varNames[j] -->
<!--         removeTermsGQ <- character() ##set as empty to begin loop -->
<!--         removeTermsNSEW <- character() -->
<!--         maxPValGQ <- 1 -->
<!--         maxPValNSEW <- 1 -->


<!--         for(i in 1:22){ ##iterate through regressions -->
<!--                 if(maxPValGQ < plim & maxPValNSEW < plim){ ##if the max p values for both GQ and NSEW are less than 0.2 then break -->
<!--                         break -->
<!--                 } -->
<!--                 print(c(maxPValGQ, maxPValNSEW)) -->
<!--                 print(c(removeTermsGQ)) -->
<!--                 print(removeTermsNSEW) -->

<!--                 ##Create new formula based on remove terms -->
<!--                 regressionFormula <- createFormula(yVariable = y, gqVars = zVarsGQ[!(zVarsGQ %in% removeTermsGQ)], nsewVars = zVarsNSEW[!(zVarsNSEW %in% removeTermsNSEW)]) -->

<!--                 ##run the regression -->
<!--                 model <- plm(regressionFormula, data = allData) -->

<!--                 ##Patterns for extracting GQ and NSEW estimates -->
<!--                 patternGQ <- "(?=.*gqDistType0-40)(?=.*postGQ(?!$))" -->
<!--                 patternNSEW <- "(?=.*nsewDistType0-40)(?=.*postNSEW(?!$))" -->

<!--                 ##Extract the estimates with max p value -->
<!--                 gqEstimates <- tidy(model) %>%  -->
<!--                         filter(str_detect(term, patternGQ)) %>%  -->
<!--                         filter(p.value == max(p.value, na.rm = T)) -->

<!--                 nsewEstimates <- tidy(model) %>%  -->
<!--                         filter(str_detect(term, patternNSEW)) %>%  -->
<!--                         filter(p.value == max(p.value, na.rm = T)) -->

<!--                 ##Identify max pvalue and terms to be removed from formula (if current value is greater than threshold) -->
<!--                 maxPValGQ <- ifelse(maxPValGQ > plim, gqEstimates$p.value, maxPValGQ) ##global assignemnt <<- -->
<!--                 maxPValNSEW <- ifelse(maxPValNSEW > plim, nsewEstimates$p.value, maxPValNSEW) ##global assignemnt <<- -->

<!--                 if(maxPValGQ > plim & maxPValGQ > maxPValNSEW){ ##update the removal term if p value is less than threshold -->
<!--                         removeTermsGQ <- c(removeTermsGQ, str_split(gqEstimates$term, ":", simplify = T)[1, 3])  -->
<!--                 } else if(maxPValNSEW > plim){ -->
<!--                         removeTermsNSEW <- c(removeTermsNSEW, str_split(nsewEstimates$term, ":", simplify = T)[1, 3])  -->
<!--                 } -->
<!--         } -->
<!--         regressionSpecs <- c(regressionSpecs, (as.character(regressionFormula)[3])) ##store the terms -->
<!-- } -->







<!-- ``` -->
